agent/requirements.txt
click>=8.1.7
rich>=13.7.0
groq>=0.9.0
ollama>=0.3.0
python-dotenv>=1.0.0
langchain-groq>=0.1.9
langchain-ollama>=0.1.3
langchain-core>=0.2.0
